src/main/webapp/app/entities/telefone/update/telefone-update.component.ts
import { Component, OnInit, inject } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize, map } from 'rxjs/operators';

import SharedModule from 'app/shared/shared.module';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';

import { IPessoa } from 'app/entities/pessoa/pessoa.model';
import { PessoaService } from 'app/entities/pessoa/service/pessoa.service';
import { ITelefone } from '../telefone.model';
import { TelefoneService } from '../service/telefone.service';
import { TelefoneFormGroup, TelefoneFormService } from './telefone-form.service';

@Component({
  standalone: true,
  selector: 'jhi-telefone-update',
  templateUrl: './telefone-update.component.html',
  imports: [SharedModule, FormsModule, ReactiveFormsModule],
})
export class TelefoneUpdateComponent implements OnInit {
  isSaving = false;
  telefone: ITelefone | null = null;

  pessoasSharedCollection: IPessoa[] = [];

  protected telefoneService = inject(TelefoneService);
  protected telefoneFormService = inject(TelefoneFormService);
  protected pessoaService = inject(PessoaService);
  protected activatedRoute = inject(ActivatedRoute);

  // eslint-disable-next-line @typescript-eslint/member-ordering
  editForm: TelefoneFormGroup = this.telefoneFormService.createTelefoneFormGroup();

  comparePessoa = (o1: IPessoa | null, o2: IPessoa | null): boolean => this.pessoaService.comparePessoa(o1, o2);

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ telefone }) => {
      this.telefone = telefone;
      if (telefone) {
        this.updateForm(telefone);
      }

      this.loadRelationshipsOptions();
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const telefone = this.telefoneFormService.getTelefone(this.editForm);
    if (telefone.id !== null) {
      this.subscribeToSaveResponse(this.telefoneService.update(telefone));
    } else {
      this.subscribeToSaveResponse(this.telefoneService.create(telefone));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<ITelefone>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(telefone: ITelefone): void {
    this.telefone = telefone;
    this.telefoneFormService.resetForm(this.editForm, telefone);

    this.pessoasSharedCollection = this.pessoaService.addPessoaToCollectionIfMissing<IPessoa>(
      this.pessoasSharedCollection,
      telefone.pessoa,
    );
  }

  protected loadRelationshipsOptions(): void {
    this.pessoaService
      .query()
      .pipe(map((res: HttpResponse<IPessoa[]>) => res.body ?? []))
      .pipe(map((pessoas: IPessoa[]) => this.pessoaService.addPessoaToCollectionIfMissing<IPessoa>(pessoas, this.telefone?.pessoa)))
      .subscribe((pessoas: IPessoa[]) => (this.pessoasSharedCollection = pessoas));
  }
}
